requirements.txt
asgiref==3.3.1
attrdict==2.0.1
autopep8==1.5.7
bleach==3.3.0
boto3==1.17.42
botocore==1.20.42
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.1.7
django-allauth==0.44.0
django-crispy-forms==1.11.0
django-doclib==1.2.2
django-environ==0.4.5
django-filter==2.4.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.1.0
django-rest-swagger==2.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
dnspython==2.1.0
docutils==0.16
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud-core==1.6.0
google-cloud-storage==1.37.0
google-crc32c==1.1.2
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
gunicorn==20.0.4
hello-pratyush1606==1.0.0
httpie==2.4.0
httplib2==0.19.1
idna==2.10
importlib-metadata==3.7.3
install==1.3.4
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
joblib==1.0.1
keyring==23.0.0
Markdown==3.3.3
MarkupSafe==1.1.1
mock==4.0.3
nltk==3.5
numpy==1.20.1
oauthlib==3.1.0
openapi-codec==1.3.2
packaging==20.9
pandas==1.2.3
pdfminer==20191125
pkginfo==1.7.0
protobuf==3.15.6
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
Pygments==2.7.4
PyJWT==1.7.1
pymongo==3.11.3
pyparsing==2.4.7
PyPDF2==1.26.0
PySocks==1.7.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
pywin32-ctypes==0.2.0
PyYAML==5.4.1
readme-renderer==29.0
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
responses==0.13.3
rfc3986==1.4.0
rsa==4.7.2
s3transfer==0.3.6
sh==1.14.1
simplejson==3.17.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tqdm==4.59.0
tweepy==3.10.0
twine==3.4.1
typing-extensions==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.4
webencodings==0.5.1
websocket-client==0.58.0
websockets==8.1
whitenoise==5.2.0
zipp==3.4.0
